# server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Fourq Distro API",
    "version": "1.0.0",
    "description": "Local listing manager for selling a vinyl record collection across eBay, Discogs, and Facebook Marketplace."
  },
  "servers": [
    { "url": "http://localhost:3000" }
  ],
...
  "paths": {
    "/api/inventory": {
      "get": {
        "tags": ["Inventory"],
        "summary": "List / filter records",
        "parameters": [
          { "name": "tier", "in": "query", "schema": { "type": "string" }, "description": "Prefix-matches the tier field" },
          { "name": "ebay_status", "in": "query", "schema": { "type": "string", "enum": ["unlisted", "listed"] } },
          { "name": "fb_status", "in": "query", "schema": { "type": "string", "enum": ["unlisted", "listed"] } },
          { "name": "discogs_status", "in": "query", "schema": { "type": "string", "enum": ["unlisted", "listed"] } },
          { "name": "search", "in": "query", "schema": { "type": "string" }, "description": "Matches artist, title, or label" },
          { "name": "sold", "in": "query", "schema": { "type": "string", "enum": ["true", "false"] } }
        ],
        "responses": {
          "200": {
            "description": "Matching records",
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Record" } } } }
          }
        }
      },
      "post": {
        "tags": ["Inventory"],
        "summary": "Create a blank record",
        "description": "Used by the photo-identification workflow — everything gets filled in after uploading photos.",
        "responses": {
          "201": {
            "description": "The newly-created blank record",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Record" } } }
          }
        }
      }
    },
    "/api/inventory/{id}": {
      "get": {
        "tags": ["Inventory"],
        "summary": "Get one record",
        "parameters": [{ "$ref": "#/components/parameters/RecordId" }],
        "responses": {
          "200": { "description": "The record", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Record" } } } },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "patch": {
        "tags": ["Inventory"],
        "summary": "Update a record",
        "description": "Only a fixed allow-list of fields can be patched: artist, title, format, year, country, label, asking_price, condition_media, condition_sleeve, matrix_number, sealed, notes, listing_extras, status, sold, discogs_sold_stats.",
        "parameters": [{ "$ref": "#/components/parameters/RecordId" }],
        "requestBody": {
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RecordPatch" } } }
        },
        "responses": {
          "200": { "description": "The updated record", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Record" } } } },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["Inventory"],
        "summary": "Delete a record",
        "parameters": [{ "$ref": "#/components/parameters/RecordId" }],
        "responses": {
          "200": { "description": "Deleted", "content": { "application/json": { "schema": { "type": "object", "properties": { "ok": { "type": "boolean" } } } } } },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/api/settings": {
      "get": {
        "tags": ["Inventory"],
        "summary": "App-wide settings",
        "responses": {
          "200": {
            "description": "Current settings",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "auto_tag_deadwax_photos": { "type": "boolean", "description": "Whether photo uploads run the local-Ollama dead wax photo classification. On by default." },
                    "auto_identify_from_photos": { "type": "boolean", "description": "Whether photo uploads run the local-Ollama artist/title/label/format/year/country extraction on a new arrival's cover/label photos. Only fills fields still blank, only while artist and title are both blank. On by default." }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Inventory"],
        "summary": "Update app-wide settings",
        "description": "Merges the given fields into the current settings.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "type": "object", "properties": { "auto_tag_deadwax_photos": { "type": "boolean" }, "auto_identify_from_photos": { "type": "boolean" } } }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated settings",
            "content": { "application/json": { "schema": { "type": "object", "properties": { "auto_tag_deadwax_photos": { "type": "boolean" }, "auto_identify_from_photos": { "type": "boolean" } } } } }
          }
        }
      }
    },
    "/api/inventory/{id}/price-suggestion": {
      "get": {
        "tags": ["Pricing & Discogs"],
        "summary": "App's own suggested price for this record",
        "description": "Grade-based, computed from a previously-looked-up discogs_price_by_grade — not a live call.",
        "parameters": [{ "$ref": "#/components/parameters/RecordId" }],
        "responses": {
          "200": {
            "description": "Suggestion",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "suggested": { "type": ["number", "null"] },
                    "reason": { "type": "string" },
                    "demand": { "type": ["string", "null"] }
                  }
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/inventory/{id}/discogs-lookup": {
      "post": {
        "tags": ["Pricing & Discogs"],
        "summary": "Find/refresh this record's Discogs match",
        "description": "Searches Discogs by artist/title if no release is matched yet (or reuses discogs_release_id if already set), then fetches price-by-grade and community stats. Re-sorts the tier if it's still in a provisional state.",
        "parameters": [{ "$ref": "#/components/parameters/RecordId" }],
        "responses": {
          "200": { "description": "The updated record", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Record" } } } },
          "404": { "description": "No matching release found on Discogs, or record not found" },
          "400": { "description": "Discogs is not configured (missing DISCOGS_TOKEN)" }
        }
      }
    },
    "/api/inventory/{id}/publish-discogs": {
      "post": {
        "tags": ["Pricing & Discogs"],
        "summary": "Publish a live Discogs marketplace listing",
        "parameters": [{ "$ref": "#/components/parameters/RecordId" }],
        "responses": {
          "200": { "description": "The updated record", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Record" } } } },
          "400": { "description": "Not configured, no matching release, missing condition, or missing price" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/inventory/{id}/unlist-discogs": {
      "post": {
        "tags": ["Pricing & Discogs"],
        "summary": "Remove the live Discogs listing",
        "parameters": [{ "$ref": "#/components/parameters/RecordId" }],
        "responses": {
          "200": { "description": "The updated record", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Record" } } } },
          "400": { "description": "This record has no active Discogs listing, or Discogs is not configured" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/inventory/{id}/listing": {
      "get": {
        "tags": ["Listing"],
        "summary": "Generate eBay / Facebook listing text",
        "parameters": [{ "$ref": "#/components/parameters/RecordId" }],
        "responses": {
          "200": {
            "description": "Generated listing text",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "coreFacts": { "type": "string" },
                    "ebayTitle": { "type": "string" },
                    "ebayWrapper": { "type": "string" },
                    "ebayDescription": { "type": "string" },
                    "fbTitle": { "type": "string" },
                    "fbWrapper": { "type": "string" },
                    "fbDescription": { "type": "string" }
                  }
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/api/inventory/{id}/photos/rotate": {
      "post": {
        "tags": ["Photos"],
        "summary": "Rotate one photo 90° increments",
        "description": "Rewrites the photo file in place and invalidates its cached hosted URL so the next publish re-uploads the corrected image.",
        "parameters": [{ "$ref": "#/components/parameters/RecordId" }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "url": { "type": "string", "example": "/photos/<record-id>/162...-front.jpg" },
                  "degrees": { "type": "number", "example": 90 }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "The updated record", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Record" } } } },
          "400": { "description": "Unknown photo url for this record" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "RecordId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string" },
        "description": "Record ID — a slug derived from artist/title/label for imported records, or a UUID for ones created via the photo-ID workflow."
      }
    },
    "responses": {
      "NotFound": {
        "description": "Record not found",
        "content": { "application/json": { "schema": { "type": "object", "properties": { "error": { "type": "string", "example": "Not found" } } } } }
      }
    },
...
  }
}